共有モジュールにpem変換を追加
